requirements.txt
aiogram~=3.4.1
python-dotenv~=1.0.1

pydantic~=2.5.3
requests~=2.31.0